TimeMeasure.h: move to include/dg/utils
