+++ tools/llvm-dda-dump.cpp
@@ -26,9 +26,10 @@
 #include "dg/llvm/DataDependence/DataDependence.h"
 #include "dg/llvm/PointerAnalysis/PointerAnalysis.h"
 
-#include "dg/tools/TimeMeasure.h"
 #include "dg/tools/llvm-slicer-opts.h"
 #include "dg/tools/llvm-slicer-utils.h"
+
+#include "dg/util/TimeMeasure.h"
 #include "dg/util/debug.h"
 
 using namespace dg;
